Refactoring : Uniformisation Structure Composants
