+++ apollo-server/composants/auth/AuthErreurs.js
@@ -1,0 +1,13 @@
+import CodinSchoolError from '../erreur'
+
+export class AuthentificationRequiseError extends CodinSchoolError {
+  constructor() {
+    super('AUTH_REQUISE', 'Cette requête requière une authentification.')
+  }
+}
+
+export class AccesInterditError extends CodinSchoolError {
+  constructor(role, interdits) {
+    super('ACCES_REFUSE', 'Droits d\'accès insuffisants.', { role, interdits })
+  }
+}
