+++ apollo-server/composants/auth/index.graphql
@@ -1,43 +1,1 @@
 #import * from "./DirectiveAcces.graphql"
-#import * from "../utilisateur/index.graphql"
-
-type Query @acces {
-	moi: Profile!
-
-	utilisateur(id: ID!): Profile
-}
-
-type Mutation {
-	inscription(inscription: Inscription!): Inscrit!
-	activer(code: String!): String!
-
-	connexion(email: String!, motDePasse: String!): Authentifie!
-
-	motDePasseOublie(email: String!): String!
-	reinitialisationMotDePasse(email: String!, code: String!, motDePasse: String!): String!
-
-	editer(modifications: EditionUtilisateur): Query! @acces
-
-	creerUtilisateur(utilisateur: Inscription): Profile!
-	@acces(interdit: [ Etudiant, Redacteur ])
-	editerUtilisateur(id: ID!, utilisateur: EditionUtilisateur): Profile!
-	@acces(interdit: [ Etudiant, Redacteur ])
-	supprimerUtilisateur(id: ID!): Profile!
-	@acces(interdit: [ Etudiant, Redacteur ])
-}
-
-"""
-Décrit ce que permet une authentification réussie.
-"""
-type Authentifie {
-	jeton: String!
-
-	moi: Profile!
-}
-
-"""
-Décrit ce qui est retourné lors d'une inscription.
-"""
-type Inscrit {
-	email: String!
-}
