+++ apollo-server/composants/exercice/index.js
@@ -1,0 +1,2 @@
+export { default as Exercice } from './ExerciceModele'
+export { default as resolvers } from './ExerciceResolvers'
