+++ apollo-server/composants/utilisateur/ProfileErreurs.js
@@ -1,17 +1,5 @@
 import CodinSchoolError from '../erreur'
 
-export class AuthentificationRequiseError extends CodinSchoolError {
-  constructor() {
-    super('AUTH_REQUISE', 'Cette requête requière une authentification.')
-  }
-}
-
-export class AccesInterditError extends CodinSchoolError {
-  constructor(role, interdits) {
-    super('ACCES_REFUSE', 'Droits d\'accès insuffisants.', { role, interdits })
-  }
-}
-
 export class IdentifiantsNonReconnusError extends CodinSchoolError {
   constructor(email) {
     super('IDENTIFIANTS_INVALIDES', 'Adresse email ou mot de passe incorrect.', { email })
