+++ apollo-server/composants/utilisateur/index.graphql
@@ -1,4 +1,30 @@
+#import * from "../auth/DirectiveAcces.graphql"
 #import * from "./Profile.graphql"
+
+type Query @acces {
+	moi: Profile!
+
+	utilisateur(id: ID!): Profile
+}
+
+type Mutation {
+	inscription(inscription: Inscription!): Inscrit!
+	activer(code: String!): String!
+
+	connexion(email: String!, motDePasse: String!): Authentifie!
+
+	motDePasseOublie(email: String!): String!
+	reinitialisationMotDePasse(email: String!, code: String!, motDePasse: String!): String!
+
+	editer(modifications: EditionUtilisateur): Query! @acces
+
+	creerUtilisateur(utilisateur: Inscription): Profile!
+	@acces(interdit: [ Etudiant, Redacteur ])
+	editerUtilisateur(id: ID!, utilisateur: EditionUtilisateur): Profile!
+	@acces(interdit: [ Etudiant, Redacteur ])
+	supprimerUtilisateur(id: ID!): Profile!
+	@acces(interdit: [ Etudiant, Redacteur ])
+}
 
 """
 Représente une demande d'inscription d'un Utilisateur.
@@ -39,4 +65,20 @@ input EditionUtilisateur {
 	siteWeb: String
 
 	avatar: String
+}
+
+"""
+Décrit ce que permet une authentification réussie.
+"""
+type Authentifie {
+	jeton: String!
+
+	moi: Profile!
+}
+
+"""
+Décrit ce qui est retourné lors d'une inscription.
+"""
+type Inscrit {
+	email: String!
 }
