+++ apollo-server/composants/utilisateur/index.js
@@ -1,9 +1,4 @@
-import Profile from './Profile'
-
-export { Profile }
-
-export const recupererParID = id => Profile.findByPk(id)
-export const recupererParEmail = email =>
-  Profile.findOne({ where: { emailPrimaire: { in: [email.toLowerCase()] } } })
-export const recupererParValidation = validation =>
-  Profile.findOne({ where: { validationInscription: { in: [validation] } } })
+export * from './ProfileErreurs'
+export * from './ProfileLogique'
+export { default as Profile } from './ProfileModele'
+export { default as resolvers } from './ProfileResolvers'
