+++ apollo-server/mocks.js
@@ -1,7 +1,6 @@
 // Enable mocking in vue.config.js with `"pluginOptions": { "enableMocks": true }`
 // Customize mocking: https://www.apollographql.com/docs/graphql-tools/mocking.html#Customizing-mocks
 
-// export default {
-//   // Mock resolvers here
-// }
-export default {}
+export default {
+  // Mock resolvers here
+}
