+++ apollo-server/resolvers.js
@@ -1,7 +1,7 @@
 import deepmerge from 'deepmerge'
 
-import Niveau from './composants/niveau/resolvers.js'
-import Exercice from './composants/exercice/resolvers.js'
-import Authentification from './composants/auth/resolvers.js'
+import Niveau from './composants/niveau/NiveauResolvers'
+import Exercice from './composants/exercice/ExerciceResolvers'
+import Utilisateur from './composants/utilisateur/ProfileResolvers'
 
-export default deepmerge.all([{ Query: {}, Mutation: {} }, Niveau, Exercice, Authentification])
+export default deepmerge.all([{ Query: {}, Mutation: {} }, Niveau, Exercice, Utilisateur])
